requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
tensorflow
streamlit
opencv-python==4.8.1.78